Add support for coalescing decimal64 ranges.
 * (M) pkg/yang/types_builtin.go
  - Add support for add() on a Number that represents as decimal64,
    rather than an integer number. This was required to support modules
    that have multiple ranges that need to be coalesced together. See
    github.com/openconfig/ygot/issues/361.
 * (R) pkg/yang/range_test.go -> pkg/yang/number_test.go
  - Rename range_test to number_test to make it more general.
  - Add unit testing for the .add() function that was previously not
    explicitly tested (other than during range testing).
