+++ pkg/yang/number_test.go
@@ -15,322 +15,227 @@
 package yang
 
 import (
-	"fmt"
 	"testing"
+
+	"github.com/google/go-cmp/cmp"
 )
 
-// DebugString returns n's internal represenatation as a string.
-func (n Number) DebugString() string {
-	return fmt.Sprintf("{%v, %d, %d}", n.Kind, n.Value, n.FractionDigits)
-}
+const (
+	useMin = -999
+	useMax = 999
+)
 
-// errToStr outputs e's error string if it is not-nil, or an empty string
-// otherwise.
-func errToStr(e error) string {
-	if e == nil {
-		return ""
+func R(a, b int64) YRange {
+	n1 := FromInt(a)
+	n2 := FromInt(b)
+	if a == useMin {
+		n1 = minNumber
 	}
-	return e.Error()
-}
-
-func TestNumberParse(t *testing.T) {
-	tests := []struct {
-		desc      string
-		numString string
-		want      Number
-		wantErr   string
-	}{{
-		desc:      "+ve int",
-		numString: "123",
-		want:      Number{Kind: Positive, Value: 123},
-	}, {
-		desc:      "-ve int",
-		numString: "-123",
-		want:      Number{Kind: Negative, Value: 123},
-	}, {
-		desc:      "+ve float",
-		numString: "123.123",
-		want:      Number{Kind: Positive, Value: 123123, FractionDigits: 3},
-	}, {
-		desc:      "+ve float, no leading 0",
-		numString: ".123",
-		want:      Number{Kind: Positive, Value: 123, FractionDigits: 3},
-	}, {
-		desc:      "+ve float, leading 0",
-		numString: "0.123",
-		want:      Number{Kind: Positive, Value: 123, FractionDigits: 3},
-	}, {
-		desc:      "-ve float, small value",
-		numString: "-0.0123",
-		want:      Number{Kind: Negative, Value: 123, FractionDigits: 4},
-	}, {
-		desc:      "+ve float, small value",
-		numString: "0.0123",
-		want:      Number{Kind: Positive, Value: 123, FractionDigits: 4},
-	}, {
-		desc:      "-ve float",
-		numString: "-123.123",
-		want:      Number{Kind: Negative, Value: 123123, FractionDigits: 3},
-	}, {
-		desc:      "bad string",
-		numString: "abc",
-		want:      Number{},
-		wantErr:   `abc is not a valid decimal number: strconv.ParseInt: parsing "abc": invalid syntax`,
-	}, {
-		desc:      "overflow ParseInt",
-		numString: "123456789123456789123456789",
-		want:      Number{},
-		wantErr:   `123456789123456789123456789 is not a valid decimal number: strconv.ParseInt: parsing "123456789123456789123456789": value out of range`,
-	}, {
-		desc:      "+ve range edge",
-		numString: "922337203685477580.7",
-		want:      Number{Kind: Positive, Value: 9223372036854775807, FractionDigits: 1},
-	}, {
-		desc:      "-ve range edge",
-		numString: "-922337203685477580.8",
-		want:      Number{Kind: Negative, Value: 9223372036854775808, FractionDigits: 1},
-	}, {
-		desc:      "overflow range +ve, frac digits 1",
-		numString: "922337203685477580.8",
-		want:      Number{},
-		wantErr:   `922337203685477580.8 is not a valid decimal number: strconv.ParseInt: parsing "9223372036854775808": value out of range`,
-	}, {
-		desc:      "overflow range -ve, frac digits 1",
-		numString: "-922337203685477580.9",
-		want:      Number{},
-		wantErr:   `-922337203685477580.9 is not a valid decimal number: strconv.ParseInt: parsing "-9223372036854775809": value out of range`,
-	}, {
-		desc:      "overflow range +ve, frac digits 18",
-		numString: "9.223372036854775808",
-		want:      Number{},
-		wantErr:   `9.223372036854775808 is not a valid decimal number: strconv.ParseInt: parsing "9223372036854775808": value out of range`,
-	}, {
-		desc:      "overflow range -ve, frac digits 18",
-		numString: "-9.223372036854775809",
-		want:      Number{},
-		wantErr:   `-9.223372036854775809 is not a valid decimal number: strconv.ParseInt: parsing "-9223372036854775809": value out of range`,
-	}, {
-		desc:      "overflow range, frac digits 19",
-		numString: "9.2233720368547758090",
-		want:      Number{},
-		wantErr:   `9.2233720368547758090 is not a valid decimal number: strconv.ParseInt: parsing "92233720368547758090": value out of range`,
-	}}
-
-	for _, tt := range tests {
-		n, err := ParseNumber(tt.numString)
-		if got, want := errToStr(err), tt.wantErr; got != want {
-			t.Errorf("%s: got error: %v, want error: %v", tt.desc, got, want)
-		}
-		if got, want := n, tt.want; tt.wantErr == "" && !got.Equal(want) {
-			t.Errorf("%s: got: %v, want: %v", tt.desc, got, want)
-		}
-		if err == nil {
-			want := tt.numString
-			if want[0] == '.' {
-				want = "0" + want
-			}
-			if got := n.String(); got != want {
-				t.Errorf("%s: got %q, want %q", tt.desc, got, want)
-			}
-		}
+	if b == useMax {
+		n2 = maxNumber
 	}
+	return YRange{n1, n2}
 }
 
-func TestNumberFromFloat(t *testing.T) {
-	tests := []struct {
-		desc string
-		num  float64
-		want Number
-	}{{
-		desc: "+ve integer",
-		num:  123,
-		want: Number{Kind: Positive, Value: 123},
-	}, {
-		desc: "-ve integer",
-		num:  -123,
-		want: Number{Kind: Negative, Value: 123},
-	}, {
-		desc: "+ve float",
-		num:  123.123,
-		want: Number{Kind: Positive, Value: 123123, FractionDigits: 3},
-	}, {
-		desc: "-ve float",
-		num:  -123.123,
-		want: Number{Kind: Negative, Value: 123123, FractionDigits: 3},
-	}, {
-		desc: "+ve max",
-		num:  float64(MaxInt64),
-		want: Number{Kind: Positive, Value: 9223372036854775808},
-	}, {
-		desc: "-ve max",
-		num:  float64(MinInt64),
-		want: Number{Kind: Negative, Value: 9223372036854775808},
-	}, {
-		desc: "+ve overflow",
-		num:  999e99,
-		want: maxNumber,
-	}, {
-		desc: "-ve overflow",
-		num:  -999e99,
-		want: minNumber,
-	}}
-
-	for _, tt := range tests {
-		n := FromFloat(tt.num)
-		if got, want := n, tt.want; !got.Equal(want) {
-			t.Errorf("%s: got: %s, want: %s", tt.desc, got.DebugString(), want.DebugString())
-		}
-
-	}
-}
-
-func TestNumberIsDecimal(t *testing.T) {
+func TestRangeEqual(t *testing.T) {
 	for x, tt := range []struct {
-		n  Number
-		ok bool
-	}{
-		{FromInt(42), false},
-		{FromInt(-41), false},
-		{minNumber, false},
-		{maxNumber, false},
-		{FromFloat(42.42), true},
-		{FromFloat(-42.42), true},
-		{Number{Kind: Positive, Value: 42}, false},
-	} {
-		ok := tt.n.IsDecimal()
-		if ok != tt.ok {
-			t.Errorf("#%d: got %v, want %v", x, ok, tt.ok)
-		}
-	}
-}
-
-func TestNumberLess(t *testing.T) {
-	for x, tt := range []struct {
-		n1, n2 Number
+		r1, r2 YangRange
 		ok     bool
 	}{
-		{FromInt(42), FromInt(42), false},
-		{FromInt(41), FromInt(42), true},
-		{FromInt(42), FromInt(41), false},
-		{FromInt(-10), FromInt(10), true},
-		{FromInt(-10), FromInt(1), true},
-		{FromInt(-10), FromInt(-1), true},
-		{FromInt(2), FromInt(-1), false},
-		{minNumber, minNumber, false},
-		{minNumber, maxNumber, true},
-		{minNumber, FromInt(42), true},
-		{minNumber, FromInt(0), true},
-		{minNumber, FromInt(-42), true},
-		{maxNumber, maxNumber, false},
-		{maxNumber, minNumber, false},
-		{maxNumber, FromInt(42), false},
-		{maxNumber, FromInt(0), false},
-		{FromInt(-42), maxNumber, true},
-		{FromInt(0), maxNumber, true},
-		{FromInt(42), maxNumber, true},
-		{FromFloat(42.42), FromFloat(42.42), false},
-		{FromFloat(41.42), FromFloat(42.42), true},
-		{FromFloat(42.42), FromFloat(41.42), false},
-		{FromFloat(42.42), FromFloat(42.421), true},
-		{FromFloat(42.1), FromFloat(42.05), false},
-		{FromFloat(41.421), FromFloat(42.42), true},
-		{FromFloat(-10.42), FromFloat(10.42), true},
-		{FromFloat(-10.42), FromFloat(1.42), true},
-		{FromFloat(-10.42), FromFloat(-1.42), true},
-		{FromFloat(-10.1), FromFloat(-10.05), true},
-		{FromFloat(2.42), FromFloat(-1.42), false},
-		{FromFloat(1234567890), FromFloat(0.123456789), false},
-		{FromFloat(-1234567890), FromFloat(0.123456789), true},
-		{minNumber, FromFloat(42.42), true},
-		{minNumber, FromFloat(0.42), true},
-		{minNumber, FromFloat(-42.42), true},
-		{maxNumber, FromFloat(42.42), false},
-		{maxNumber, FromFloat(0.42), false},
-		{FromFloat(-42.42), maxNumber, true},
-		{FromFloat(0.42), maxNumber, true},
-		{FromFloat(42.42), maxNumber, true},
-		{FromInt(42), FromFloat(42), false},
-		{FromInt(41), FromFloat(42), true},
-		{FromInt(42), FromFloat(42.42), true},
-		{FromInt(-42), FromFloat(-42), false},
-		{FromInt(-42), FromFloat(-41), true},
-		{FromInt(-42), FromFloat(-42.42), false},
-		{Number{Kind: Positive, Value: 9223372036854775807, FractionDigits: 0},
-			Number{Kind: Positive, Value: 9223372036854775807, FractionDigits: 18}, false},
-		{Number{Kind: Negative, Value: 9223372036854775808, FractionDigits: 0},
-			Number{Kind: Positive, Value: 9223372036854775808, FractionDigits: 18}, true},
-		{Number{Kind: Positive, Value: 9223372036854775807, FractionDigits: 1},
-			Number{Kind: Positive, Value: 9223372036854775807, FractionDigits: 18}, false},
-		{Number{Kind: Negative, Value: 9223372036854775808, FractionDigits: 1},
-			Number{Kind: Positive, Value: 9223372036854775808, FractionDigits: 18}, true},
+		{ok: true},                          // empty range contained in empty range
+		{r1: YangRange{R(1, 2)}, ok: false}, // empty range contained in range
+		{r2: YangRange{R(1, 2)}, ok: false}, // range contained in empty range
+		{
+			YangRange{R(1, 2)},
+			YangRange{R(1, 2)},
+			true,
+		},
+		{
+			YangRange{R(1, 3)},
+			YangRange{R(1, 2)},
+			false,
+		},
+		{
+			YangRange{R(1, 2), R(4, 5)},
+			YangRange{R(1, 2), R(4, 5)},
+			true,
+		},
+		{
+			YangRange{R(1, 2), R(4, 6)},
+			YangRange{R(1, 2), R(4, 5)},
+			false,
+		},
+		{
+			YangRange{R(1, 2)},
+			YangRange{R(1, 2), R(4, 5)},
+			false,
+		},
+		{
+			YangRange{R(1, 2), R(4, 5)},
+			YangRange{R(1, 2)},
+			false,
+		},
 	} {
-		ok := tt.n1.Less(tt.n2)
-		if ok != tt.ok {
+		if ok := tt.r1.Equal(tt.r2); ok != tt.ok {
 			t.Errorf("#%d: got %v, want %v", x, ok, tt.ok)
 		}
 	}
 }
 
-func TestNumberEqual(t *testing.T) {
+func TestRangeContains(t *testing.T) {
 	for x, tt := range []struct {
-		n1, n2 Number
+		r1, r2 YangRange
 		ok     bool
 	}{
-		{FromInt(42), FromInt(42), true},
-		{FromInt(41), FromInt(42), false},
-		{FromInt(42), FromInt(41), false},
-		{FromInt(-10), FromInt(-10), true},
-		{FromInt(-10), FromInt(1), false},
-		{FromInt(0), FromInt(0), true},
-		{minNumber, minNumber, true},
-		{minNumber, maxNumber, false},
-		{minNumber, FromInt(42), false},
-		{minNumber, FromInt(0), false},
-		{minNumber, FromInt(-42), false},
-		{maxNumber, maxNumber, true},
-		{maxNumber, minNumber, false},
-		{maxNumber, FromInt(42), false},
-		{maxNumber, FromInt(0), false},
-		{FromInt(-42), maxNumber, false},
-		{FromInt(0), maxNumber, false},
-		{FromInt(42), maxNumber, false},
-		{FromFloat(42.42), FromFloat(42.42), true},
-		{FromFloat(41.42), FromFloat(42.42), false},
-		{FromFloat(-10.42), FromFloat(10.42), false},
-		{FromFloat(-10.42), FromFloat(-10.42), true},
-		{FromFloat(-10.42), FromFloat(1.42), false},
-		{FromFloat(-10.42), FromFloat(-1.42), false},
-		{FromFloat(2.42), FromFloat(-1.42), false},
-		{minNumber, FromFloat(42.42), false},
-		{minNumber, FromFloat(0.42), false},
-		{minNumber, FromFloat(-42.42), false},
-		{maxNumber, FromFloat(42.42), false},
-		{maxNumber, FromFloat(0.42), false},
-		{FromFloat(-42.42), maxNumber, false},
-		{FromFloat(0.42), maxNumber, false},
-		{FromFloat(42.42), maxNumber, false},
+		{ok: true},
+		{r1: YangRange{R(1, 2)}, ok: true},
+		{r2: YangRange{R(1, 2)}, ok: true},
+		{
+			r1: YangRange{R(1, 2)},
+			r2: YangRange{R(1, 2)},
+			ok: true,
+		},
+		{
+			r1: YangRange{R(1, 5)},
+			r2: YangRange{R(2, 3)},
+			ok: true,
+		},
+		{
+			r1: YangRange{R(2, 3)},
+			r2: YangRange{R(1, 5)},
+			ok: false,
+		},
+		{
+			r1: YangRange{R(1, 10)},
+			r2: YangRange{R(1, 2), R(4, 5), R(7, 10)},
+			ok: true,
+		},
+		{
+			r1: YangRange{R(1, 10)},
+			r2: YangRange{R(1, 2), R(7, 11)},
+			ok: false,
+		},
+		{
+			r1: YangRange{R(1, 9), R(11, 19), R(21, 29)},
+			r2: YangRange{R(23, 25)},
+			ok: true,
+		},
+		{
+			r1: YangRange{R(1, 9), R(11, 19), R(21, 29)},
+			r2: YangRange{R(23, 23)},
+			ok: true,
+		},
+		{
+			r1: YangRange{R(1, 9), R(11, 19), R(21, 29)},
+			r2: YangRange{R(20, 20)},
+			ok: false,
+		},
+		{
+			r1: YangRange{R(1, 10)},
+			r2: YangRange{R(useMin, useMax)},
+			ok: true,
+		},
+		{
+			r1: YangRange{R(useMin, useMax)},
+			r2: YangRange{R(1, 10)},
+			ok: true,
+		},
+		{
+			r1: YangRange{R(1024, 65535)},
+			r2: YangRange{R(useMin, 4096), R(5120, useMax)},
+			ok: true,
+		},
+		{
+			r1: YangRange{R(1024, 65535)},
+			r2: YangRange{R(-999999, 4096), R(5120, useMax)},
+			ok: false,
+		},
+		{
+			r1: YangRange{R(1024, 65535)},
+			r2: YangRange{R(useMin, 4096), R(5120, 999999)},
+			ok: false,
+		},
 	} {
-		ok := tt.n1.Equal(tt.n2)
-		if ok != tt.ok {
+		if ok := tt.r1.Contains(tt.r2); ok != tt.ok {
 			t.Errorf("#%d: got %v, want %v", x, ok, tt.ok)
 		}
 	}
 }
 
-func TestNumberAdd(t *testing.T) {
+func TestCoalesce(t *testing.T) {
 	for x, tt := range []struct {
-		in  Number
-		add uint64
-		out Number
+		in, out YangRange
 	}{
-		{Number{Positive, 0, 0}, 1, Number{Positive, 1, 0}},
-		{Number{Negative, 1, 0}, 1, Number{Positive, 0, 0}},
-		{Number{Positive, 5, 0}, 12, Number{Positive, 17, 0}},
-		{Number{Negative, 3, 0}, 10, Number{Positive, 7, 0}},
+		{},
+		{YangRange{R(1, 4)}, YangRange{R(1, 4)}},
+		{YangRange{R(1, 2), R(3, 4)}, YangRange{R(1, 4)}},
+		{YangRange{R(1, 2), R(2, 4)}, YangRange{R(1, 4)}},
+		{YangRange{R(1, 2), R(4, 5)}, YangRange{R(1, 2), R(4, 5)}},
+		{YangRange{R(1, 3), R(2, 5)}, YangRange{R(1, 5)}},
+		{YangRange{R(1, 10), R(2, 5)}, YangRange{R(1, 10)}},
+		{YangRange{R(1, 10), R(1, 2), R(4, 5), R(7, 8)}, YangRange{R(1, 10)}},
 	} {
-		out := tt.in.add(tt.add)
+		out := coalesce(tt.in)
 		if !out.Equal(tt.out) {
 			t.Errorf("#%d: got %v, want %v", x, out, tt.out)
 		}
+	}
+}
+
+func TestAdd(t *testing.T) {
+	tests := []struct {
+		desc  string
+		inVal Number
+		inAdd uint64
+		want  Number
+	}{{
+		desc:  "add one to integer",
+		inVal: FromInt(1),
+		inAdd: 1,
+		want:  FromInt(2),
+	}, {
+		desc:  "add one to decimal64",
+		inVal: FromFloat(1.0),
+		inAdd: 1,
+		want:  FromFloat(2.0),
+	}, {
+		desc:  "negative int becomes positive",
+		inVal: FromInt(-2),
+		inAdd: 3,
+		want:  FromInt(1),
+	}, {
+		desc:  "negative int stays negative",
+		inVal: FromInt(-3),
+		inAdd: 1,
+		want:  FromInt(-2),
+	}, {
+		desc:  "negative decimal becomes positive",
+		inVal: FromFloat(-2),
+		inAdd: 3,
+		want:  FromFloat(1.0),
+	}, {
+		desc:  "negative decimal stays negative",
+		inVal: FromFloat(-42),
+		inAdd: 41,
+		want:  FromFloat(-1.0),
+	}, {
+		desc:  "explicitly set fraction digits",
+		inVal: Number{Value: 10000, FractionDigits: 5},
+		inAdd: 1,
+		want:  Number{Value: 110000, FractionDigits: 5},
+	}, {
+		desc:  "explicitly set fraction digits - negative",
+		inVal: Number{Value: 0, FractionDigits: 10},
+		inAdd: 42,
+		want:  FromFloat(42),
+	}}
+
+	for _, tt := range tests {
+		t.Run(tt.desc, func(t *testing.T) {
+			got := tt.inVal.add(tt.inAdd)
+			if !cmp.Equal(got, tt.want) {
+				t.Fatalf("did get expected result, got: %s, want: %s", got.String(), tt.want.String())
+			}
+		})
 	}
 }
